+++ mau_mau/sim.py
@@ -1,7 +1,6 @@
 import logging
 
-from mau_mau import cardroom
-from mau_mau.subjects import Croupier
+from mau_mau import concepts, subjects, objects
 
 log = logging.getLogger(__name__)
 
@@ -15,15 +14,16 @@ def play_game(rulesOfTheGame, players):
 
 
 def setup_game(rules, playerSeed):
-    croupier = Croupier()
+    croupier = subjects.Croupier()
+    table = objects.Table()
+
     croupier.fetch_fresh_deck_of_cards()
-    table = cardroom.Table()
     croupier.invite(playerSeed, table)
     croupier.check_setup(table.players, rules.cardsPerPlayer)
     croupier.set_table(table, rules)
     for player in table.players:
         croupier.deal_fresh_hand(player, table.stock, rules.cardsPerPlayer)
     croupier.check_table(table)
-    game = cardroom.Game(table)
+    game = concepts.Game(table)
     log.debug("Start new game: %s", game)
     return game
