+++ tests/test_pile.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mau_mau.cardroom import Stock, Card, Waste
+from mau_mau.objects import Stock, Waste, Card
 
 
 def test_fetch_card_from_empty_stock_raises():
